pyproject.toml
[project]
name = "hunau-teaching-evaluation-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feapder[render]>=1.8.9",
    "setuptools>=70.0",
    "python-dotenv>=1.0",
]
